Upgrade to Rails 7.2
Note we're removing the `secret_key_base` key from the test and
development secrets, since it's ignored in these environments. Quoting
from the Rails documentation [1]:

> In development and test, this is randomly generated and stored in a
> temporary file in <tt>tmp/local_secret.txt</tt>.

Due to this, in **development**, existing cookies won't be used after
upgrading. If this is important for you (for example, if you're
constantly switching branches with Consul Democracy 2.4 and Consul
Democracy 2.5 in order to change the upgrade), **on your local
machine**, delete your `secret_key_base` references for the
**development** environment in your `config/secrets.yml`.

Other minor modifications from the code the `rails app:update` generates
include removing the `APP_NAME` and puma-dev references in `bin/setup`
(they were added in Rails 7.2 but removed once again in Rails 8.0) and
commenting the `config.autoload_lib` option (we don't autoload the
`lib/` folder but use `app/lib/` instead).

[1] https://api.rubyonrails.org/v7.2.0/classes/Rails/Application.html#method-i-secret_key_base