Check flash messages in SDG management
Note this did not result in flaky tests, since the tests continue by
clicking the  "Marked as reviewed" link, which isn't present after the
request has finished. However, we're adding an expectation so we don't
have to think whether the test could be flaky and because this way we're
also testing the user experience; it would be strange for a user if they
were redirected to a page without a flash message.