Fix wrong expectation in information texts test
The expectation that there isn't a "Español" wasn't updated in commit
f7417d647, when the tests were updated after switching the user
interface from a list of links to a select field.

We're also adding an additional expectation after the `visit` to make
sure the request has finished before selecting English from the selector
(which probably doesn't make sense because it's the only option, but
changing it is out of the scope of this commit).