Don't set seconds in poll starts_at/ends_at demo seeds
We were already doing it for dev seeds since commit 855fedc02.

The reason is that it was impossible to update the poll because the
start date can't be changed once the poll has started, and we were even
storing the nanoseconds in the database. Since the datetime fields in a
form don't send the nanoseconds, we were getting a validation error
because the date we were sending using the form was not exactly the same
as the date in the database.