Vendor Foundation form builder to remove gem dependency
The "foundation_rails_helper" gem is no longer maintained and is
incompatible with Rails 7.1. To avoid blocking the upgrade, we've vendored
the necessary parts of the gem's FormBuilder into "vendor" folder.

This commit includes the vendor/foundation_rails_helper/form_builder.rb as
a copy of the original FormBuilder class. Includes too the
vendor/foundation_rails_helper/size_class_calculator.rb that is required by
the builder, and finally includes a vendor/foundation_rails_helper/configuration.rb
to mantain compatibility with auto_labels and button_class variables, that
are used in the original builder.

The gem has been removed from the Gemfile and replaced with this vendored
fallback. This workaround is safe to remove once legacy Foundation CSS
support is dropped.

All vendored code retains the original MIT license and attribution.