Unify poster preview with poster PDF styles
Since commit c5103d30250, the styles from application.scss no longer apply
to app/views/dashboard/poster/index.pdf.erb.

To recover the text-center alignment, we add the rule in dashboard.scss.

Note that we also remove text-center from the h2, since it wasn't applied
and there's no need to recover it because it doesn't exist in the preview
either.