Keep SyslogIdentifier=puma for compatibility
We keep using SyslogIdentifier=puma instead of the default
value from capistrano3-puma 6.0.0, which is based on
puma_service_unit_name (usually "<app>_puma_<env>").

This avoids introducing a new config variable in the installer
and is enough for our current use case, where we don't run
multiple environments on the same server.